Extract a `Table` component
