+++ app/javascript/components/Header.js
@@ -4,10 +4,14 @@ import styled from "styled-components"
 import logo from "../images/logo.png"
 
 const Header = () => (
-  <div>
+  <Layout>
     <Logo src={logo} alt="Nimbus"/>
-  </div>
+  </Layout>
 )
+
+const Layout = styled.div`
+  height: 4rem;
+`
 
 const Logo = styled.img`
   height: 2rem;
